requirements.txt
alembic==1.12.0
blinker==1.8.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
Flask==2.3.3
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.3
Flask-Mail==0.9.1
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.0.5
greenlet==3.1.1
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==2.1.5
marshmallow==3.22.0
packaging==25.0
psycopg2-binary==2.9.7
PyJWT==2.9.0
python-dotenv==1.0.0
requests==2.31.0
SQLAlchemy==2.0.41
typing_extensions==4.13.2
urllib3==2.2.3
Werkzeug==2.3.7
zipp==3.20.2
